+++ opentube/extras.py
@@ -1,5 +1,3 @@
-from typing import List, Optional
-
 from .https import (
     _get_trending_learning_videos,
     trending_feeds,
@@ -16,31 +14,31 @@ from .utils import dup_filter
 class Extras:
 
     @staticmethod
-    def trending_videos() -> Optional[List[str]]:
+    def trending_videos():
         data = Patterns.video_id.findall(trending_videos())
         return dup_filter(data) if data else None
 
     @staticmethod
-    def music_videos() -> Optional[List[str]]:
+    def music_videos():
         data = Patterns.video_id.findall(trending_songs())
         return dup_filter(data) if data else None
 
     @staticmethod
-    def gaming_videos() -> Optional[List[str]]:
+    def gaming_videos():
         return dup_filter(Patterns.video_id.findall(trending_games()))
 
     @staticmethod
-    def news_videos() -> Optional[List[str]]:
+    def news_videos():
         return dup_filter(Patterns.video_id.findall(trending_feeds()))
 
     @staticmethod
-    def live_videos() -> Optional[List[str]]:
+    def live_videos():
         return dup_filter(Patterns.video_id.findall(trending_streams()))
 
     @staticmethod
-    def educational_videos() -> Optional[List[str]]:
+    def educational_videos():
         return dup_filter(Patterns.video_id.findall(_get_trending_learning_videos()))
 
     @staticmethod
-    def sport_videos() -> Optional[List[str]]:
+    def sport_videos():
         return dup_filter(Patterns.video_id.findall(trending_sports()))