+++ opentube/playlist.py
@@ -1,5 +1,4 @@
 import re
-from typing import Any, Dict
 
 from .https import playlist_data
 from .patterns import _PlaylistPatterns as Patterns
@@ -9,7 +8,7 @@ from .utils import dup_filter
 
 class Playlist:
 
-    def __init__(self, playlist_id: str):
+    def __init__(self, playlist_id):
         """
         Represents a YouTube Playlist
 
@@ -18,21 +17,21 @@ class Playlist:
         playlist_id : str
             The id or url of the playlist
         """
-        pattern = re.compile("=(.+?)$|^PL(.+?)$")
+        pattern = re.compile('=(.+?)$|^PL(.+?)$')
         match = pattern.search(playlist_id)
         if not match:
-            raise ValueError(f"Invalid playlist id: {playlist_id}")
+            raise ValueError('Invalid playlist id: %s' % playlist_id)
         if match.group(1):
             self.id = match.group(1)
         elif match.group(2):
-            self.id = "PL" + match.group(2)
+            self.id = 'PL' + match.group(2)
         self._playlist_data = playlist_data(self.id)
 
     def __repr__(self):
-        return f"<Playlist {self.id}>"
+        return '<Playlist %s>' % self.id
 
     @property
-    def metadata(self) -> Dict[str, Any]:
+    def metadata(self):
         """
         Fetches playlist metadata in a dict format
 
@@ -50,10 +49,10 @@ class Playlist:
         ext = collect(lambda x: x.findall(self._playlist_data) or None, patterns)
         data = [e[0] if e else None for e in ext]
         return {
-            "id": self.id,
-            "url": "https://www.youtube.com/playlist?list=" + self.id,
-            "name": data[0] if data else None,
-            "video_count": data[1] if data else None,
-            "thumbnail": data[2] if data else None,
-            "videos": dup_filter(ext[3]),
+            'id': self.id,
+            'url': 'https://www.youtube.com/playlist?list=' + self.id,
+            'name': data[0] if data else None,
+            'video_count': data[1] if data else None,
+            'thumbnail': data[2] if data else None,
+            'videos': dup_filter(ext[3]),
         }